refactor(test): move fixtures from service/pkg/ to test/fixtures module
Fixtures are test-only code and should not be part of the service's
public API surface. Move them to a dedicated test/fixtures module with
its own go.mod, importable by test/integration via replace directive.

- service/pkg/testutil/fixtures/ → test/fixtures/ (own go.mod)
- Extract provisionCmd to service/cmd/provision.go (provisionKeycloak needs it)
- Delete service/cmd/provisionFixtures.go from service binary
- Create test/integration/cmd/provision-fixtures/ as standalone tool
- Update all 15 integration test imports and YAML load path
