+++ test/integration/policy/registered_resources_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/opentdf/platform/protocol/go/policy/registeredresources"
 	"github.com/opentdf/platform/protocol/go/policy/unsafe"
 	"github.com/opentdf/platform/service/pkg/db"
-	"github.com/opentdf/platform/service/pkg/testutil/fixtures"
+	"github.com/opentdf/platform/test/fixtures"
 	"github.com/opentdf/platform/service/policy/actions"
 	"github.com/stretchr/testify/suite"
 	"google.golang.org/protobuf/proto"
